ModuleLoader: Don't err on extra syscalls in modules.
If a module exported system calls whose signatures were not in the set of
allowed signatures, loading that module failed. Now we only check whether
the system calls we actually need are there, instead of iterating over the
whole list and erring if any of the system calls defined by the module are
not required.